Make highlighter more generic again
