+++ studio/include/studio/syntax.hpp
@@ -28,10 +28,45 @@ class Syntax : public QSyntaxHighlighter
     Q_OBJECT
 public:
     Syntax(QTextDocument* doc) : QSyntaxHighlighter(doc) {}
-    virtual void setKeywords(QStringList kws)=0;
+    void setKeywords(QStringList kws);
+
+    using MatchInfo = std::pair<int, QString>;
+    class BlockData : public QTextBlockUserData {
+    public:
+        std::list<MatchInfo> data;
+    };
 
 public slots:
     virtual void onCursorMoved(QPlainTextEdit* text)=0;
+
+protected:
+    /*  Overrides highlightBlock in QSyntaxHighlighter */
+    void highlightBlock(const QString& text) override;
+
+    /*  Structure to use as a rule when highlighting    */
+    struct Rule {
+        Rule() {}
+        Rule(QString r, QTextCharFormat f, int si=0, int so=0, int capture=0)
+            : regex(QRegularExpression(r)), format(f), state_in(si), state_out(so),
+              capture(capture)
+        { /* Nothing to do here */ }
+
+        QRegularExpression regex;
+        QTextCharFormat format;
+        int state_in, state_out;
+        int capture;
+    };
+
+    /*  General-purpose rules */
+    QList<Rule> rules;
+
+    /* We use a special rule for language keywords, to avoid having one rule
+     * per keyword, which would slow things down by a lot. */
+    Rule m_keyword;
+    QSet<QString> m_keywords;
 };
+
+
+////////////////////////////////////////////////////////////////////////////////
 
 }   // namespace Studio
