+++ studio/src/guile/syntax.cpp
@@ -25,16 +25,6 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 ////////////////////////////////////////////////////////////////////////////////
 
-typedef std::pair<int, QString> MatchInfo;
-
-class BlockData : public QTextBlockUserData
-{
-public:
-    std::list<MatchInfo> data;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
 namespace Studio {
 namespace Guile {
 
@@ -231,83 +221,6 @@ void Syntax::onCursorMoved(QPlainTextEdit* text)
         }
     }
     text->setExtraSelections(selections);
-}
-
-void Syntax::highlightBlock(const QString& text)
-{
-    int offset = 0;
-    int state = previousBlockState();
-
-    auto data = new BlockData;
-
-    while (offset <= text.length())
-    {
-        Rule rule;
-        QRegularExpressionMatch match;
-
-        // First, check for a match against a language keyword by finding
-        // a keyword-shaped blob of text, then comparing it against the map.
-        auto m = m_keyword.regex.match(text, offset);
-        if (m.hasMatch()) {
-            if (m_keywords.contains(m.captured(0))) {
-                match = m;
-                rule = m_keyword;
-            }
-        }
-
-        // Iterate over every rule, picking out the first match
-        for (auto r : rules)
-        {
-            if (r.state_in != state)
-            {
-                continue;
-            }
-
-            auto m = r.regex.match(text, offset);
-            if (!m.hasMatch())
-            {
-                continue;
-            }
-
-            if (!match.hasMatch() ||
-                m.capturedStart(r.capture) < match.capturedStart(rule.capture) ||
-                (m.capturedStart(r.capture) <= match.capturedStart(rule.capture) &&
-                 m.capturedEnd(r.capture) > match.capturedEnd(rule.capture)))
-            {
-                match = m;
-                rule = r;
-            }
-        }
-
-        if (!match.isValid() || !match.hasMatch())
-        {
-            offset++;
-            continue;
-        }
-
-        const auto start = match.capturedStart(rule.capture);
-        const auto length = match.capturedLength(rule.capture);
-
-        // Otherwise we end up in an infinite loop
-        assert(length > 0);
-        setFormat(start, length, rule.format);
-        offset = start + length;
-        state = rule.state_out;
-
-        for (int i=1; i <= match.lastCapturedIndex(); ++i)
-        {
-            data->data.push_back({match.capturedStart(i), match.captured(i)});
-        }
-    }
-
-    setCurrentBlockState(state);
-    setCurrentBlockUserData(data);
-}
-
-void Syntax::setKeywords(QStringList kws) {
-    for (auto k: kws) {
-        m_keywords.insert(k);
-    }
 }
 
 }   // namespace Guile
